Rename all mentions
