+++ framework/py/flwr/serverapp/__init__.py
@@ -15,8 +15,8 @@
 """Public Flower ServerApp APIs."""
 
 
-from flwr.server.grid import Grid
 from flwr.server.server_app import ServerApp
+from flwr.serverapp.grid import Grid
 
 from . import strategy
